Monetary variables from Double to BigDecimal data type (#262)

+++ server/api/src/main/java/com/twb/pokerapp/dto/roundpot/RoundPotDTO.java
@@ -3,6 +3,7 @@ package com.twb.pokerapp.dto.roundpot;
 import com.twb.pokerapp.dto.playersession.PlayerSessionDTO;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -10,7 +11,7 @@ import java.util.UUID;
 @Data
 public class RoundPotDTO {
     private UUID id;
-    private Double potAmount;
+    private BigDecimal potAmount;
     private Integer potIndex;
     private List<PlayerSessionDTO> eligiblePlayers = new ArrayList<>();
 }
